% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chart_daily}
\alias{chart_daily}
\title{Spotify chart days}
\usage{
chart_daily()
}
\value{
A data frame with the days
}
\description{
This function gets available days for the charts
}
\examples{

chart_daily()

}

