% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chart_regions}
\alias{chart_regions}
\title{Spotify chart regions}
\usage{
chart_regions()
}
\value{
A data frame with the regions
}
\description{
This function gets available regions for Spotify charts
}
\examples{

chart_regions()

}

