% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data.R
\name{chart_top200_daily}
\alias{chart_top200_daily}
\title{Spotify Top 200 chart - daily}
\usage{
chart_top200_daily(region = "global", days = "latest")
}
\arguments{
\item{region}{A region defined for the chart. See \code{\link{chart_regions}} for available regions.}

\item{days}{A date defining the day for the chart. See \code{\link{chart_daily}} for available days.}
}
\value{
A data frame with the chart data
}
\description{
This function collects the 200 most popular tracks on Spotify.
}
\examples{
spotifycharts::chart_top200_daily()

spotifycharts::chart_top200_daily(region = "dk")

spotifycharts::chart_top200_daily(region = "us", days = "2016-11-03")

}

