% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data.R
\name{chart_top200_weekly}
\alias{chart_top200_weekly}
\title{Spotify Top 200 chart - weekly}
\usage{
chart_top200_weekly(region = "global", weekdates = "latest")
}
\arguments{
\item{region}{A region defined for the chart. See \code{\link{chart_regions}} for available regions.}

\item{weekdates}{A date defining the week for the chart. See \code{\link{chart_weekly_top200}} for available weeks.}
}
\value{
A data frame with the chart data
}
\description{
This function collects the 200 most popular tracks on Spotify.
}
\examples{
spotifycharts::chart_top200_weekly()

spotifycharts::chart_top200_weekly(region = "dk")

spotifycharts::chart_top200_weekly(region = "us", weekdates = "2016-09-30--2016-10-07")

}

