% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_data.R
\name{chart_viral50_weekly}
\alias{chart_viral50_weekly}
\title{Spotify Viral 50 chart - weekly}
\usage{
chart_viral50_weekly(region = "global", weekdates = "latest")
}
\arguments{
\item{region}{A region defined for the chart. See \code{\link{chart_regions}} for available regions.}

\item{weekdates}{A date defining the week for the chart. See \code{\link{chart_weekly_viral50}} for available weeks.}
}
\value{
A data frame with the chart data
}
\description{
This function collects the 50 most viral tracks on Spotify.
}
\examples{
spotifycharts::chart_viral50_weekly()

spotifycharts::chart_viral50_weekly(region = "dk")

spotifycharts::chart_viral50_weekly(region = "us", weekdates = "2016-10-27--2016-10-27")

}

