% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chart_weekly_top200}
\alias{chart_weekly_top200}
\title{Spotify Top 200 chart weeks}
\usage{
chart_weekly_top200()
}
\value{
A data frame with the weeks
}
\description{
This function gets available weeks for the Spotify Top 200 chart
}
\examples{

chart_weekly_top200()

}

