% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chart_weekly_viral50}
\alias{chart_weekly_viral50}
\title{Spotify Viral 50 chart weeks}
\usage{
chart_weekly_viral50()
}
\value{
A data frame with the weeks
}
\description{
This function gets available weeks for the Spotify Viral 50 chart
}
\examples{

chart_weekly_viral50()

}

